+++ packages/expo-manifests/ios/EXManifests/Manifest.swift
@@ -301,6 +301,16 @@ public class Manifest: NSObject {
     return supportsRTL
   }
 
+  public func forcesRTL() -> Bool {
+    guard let expoClientConfigRootObject = expoClientConfigRootObject(),
+      let extra: [String: Any]? = expoClientConfigRootObject.optionalValue(forKey: "extra"),
+      let forcesRTL: Bool = extra?.optionalValue(forKey: "forcesRTL") else {
+      return false
+    }
+
+    return forcesRTL
+  }
+
   public func jsEngine() -> String {
     let jsEngine = expoClientConfigRootObject().let { it in
       Manifest.string(fromManifest: it, atPaths: [